[user] create login profile

+++ models/base.go
@@ -5,7 +5,7 @@ import (
 )
 
 type Base struct {
-	Id        int       `json:"id" sql:",pk"`
+	Id        int64     `json:"id" sql:",pk"`
 	CreatedAt time.Time `json:"-"`
 	UpdatedAt time.Time `json:"-"`
 }
